TSP-387 Fixed Incorrect type of referrer id

+++ src/main/java/com/github/sportbot/dto/RegistrationRequest.java
@@ -14,6 +14,6 @@ public record RegistrationRequest(
         String language,
         Sex sex,
         Integer age,
-        Integer referrerTelegramId,
+        Long referrerTelegramId,
         LocalTime remindTime
 ) {}
